requirements.txt
asgiref==3.7.2
boto3==1.34.59
botocore==1.34.59
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-s3-storage==0.15.0
djangorestframework==3.14.0
idna==3.7
jmespath==1.0.1
lxml==5.1.0
mutagen==1.47.0
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dateutil==2.9.0.post0
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
